requirements.txt
proxmoxer>=1.3.0
PyYAML>=6.0
requests>=2.31.0